Suggest `ulauncher-toggle` when dbus name is taken
This adds a log when executing `ulauncher` with existing process already running that suggests user to use the `ulauncher-toggle` script instead of the full `ulauncher` which is about 250ms slower. This is especially useful for people using Wayland compositors who do their keybindings in their compositor settings and can easily be mistaken to put `ulauncher` as the executable for menu keybind instead of the faster toggle.